dags/requirements.txt
apache-airflow-providers-slack
apache-airflow-providers-snowflake